backend/requirements.txt
djangorestframework==3.15.0
